后台bd_psndoc新增人员类别字段，前台对应更新

+++ src/main/java/cn/wxd/services/psndoc/search/Delete.java
@@ -2,6 +2,7 @@ package cn.wxd.services.psndoc.search;
 
 import cn.wxd.DAO.BaseDAO;
 import cn.wxd.DAO.handler.MapListHandler;
+import cn.wxd.DAO.handler.StringHandler;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -22,18 +23,25 @@ public class Delete extends HttpServlet {
         //数据库执行返回消息
         Map<String,String> result =  new HashMap();
         StringBuffer sql = new StringBuffer();
+        //查询该人员是否用户
+        sql.append(" select PSNTYPE from BD_PSNDOC where id = '"+ID+"'");
+        String PSNTYPE = (String) new BaseDAO().executeQuarry(sql.toString(),new StringHandler());
+        sql.delete(0,sql.length());
         //删除BD_PSNDOC中的数据
         sql.append(" DELETE FROM BD_PSNDOC WHERE ID = '"+ ID+"'");
         result.put("BD_PSNDOC",new BaseDAO().executeUpdate(sql.toString()));
         sql.delete(0,sql.length());
-        //删除SM_USER的数据
-        sql.append(" DELETE FROM SM_USER WHERE PK_PSNDOC = 'ry"+ ID+"'");
-        result.put("SM_USER",new BaseDAO().executeUpdate(sql.toString()));
+        //先设置为成功，若失败则覆盖成功的值
+        result.put("SM_USER","1");
+        //如果用户类型是用户，则删除SM_USER的数据
+        if(PSNTYPE.equals("01")){
+            sql.append(" DELETE FROM SM_USER WHERE PK_PSNDOC = 'ry"+ ID+"'");
+            result.put("SM_USER",new BaseDAO().executeUpdate(sql.toString()));
+        }
         if(result.get("BD_PSNDOC").equals("1")&&result.get("SM_USER").equals("1")){
             resp.getWriter().write("1");
         }else {
             resp.getWriter().write("0");
         }
-
     }
 }
